Source-Commit: 19a6403b0d7e49ae631e346a110d098938957acc
Publish-Parent: 4032c14f103a908dad70323c27e671bc5c63aec6

+++ src/BASE/WIDGET.cpp
@@ -101,9 +101,7 @@ MessageDispatchResult widget::Main(tag_message& message) {
                 case WIDGET_COMMAND_CLEAR_FLAGS:
                     if (message.payload.widget.id == m_id) {
                         i16 flags =
-                            static_cast<i16>(
-                                message.payload.widget.data.value & WIDGET_FLAG_MASK
-                            );
+                            message.payload.widget.data.value & WIDGET_FLAG_MASK;
                         if (message.payload.widget.data.value == (WIDGET_COMMAND_DIMMED)) {
                             flags = WIDGET_FLAG_DIMMED;
                             m_flags &= ~flags;
